+++ src/irvine/oeis/a013/A013980.java
@@ -17,11 +17,11 @@ public class A013980 extends Sequence3 {
     ++mN;
     Z res = Functions.FIBONACCI.z(3L * mN - 3);
     for (int k = 2; k < mN; ++k) {
-      final Z t = Functions.FIBONACCI.z(3L * k - 4).multiply(Functions.CATALAN.z((long) (mN - k)));
+      final Z t = Functions.FIBONACCI.z(3L * k - 4).multiply(Functions.CATALAN.z(mN - k));
       res = res.subtract(t);
     }
     res = res.multiply(5);
-    res = res.add(Functions.CATALAN.z((long) (mN - 1)));
+    res = res.add(Functions.CATALAN.z(mN - 1));
     res = res.subtract(Functions.FIBONACCI.z(2L * mN - 1).multiply2());
     res = res.subtract(Functions.FIBONACCI.z(2L * mN - 3).multiply2());
     res = res.add(Functions.FIBONACCI.z(mN));
