+++ src/irvine/oeis/a022/A022290.java
@@ -13,7 +13,7 @@ public class A022290 extends Sequence0 {
   private static final Z[] FIBONACCI = new Z[Long.SIZE];
   static {
     for (int k = 0; k < FIBONACCI.length; ++k) {
-      FIBONACCI[k] = Functions.FIBONACCI.z((long) (k + 2));
+      FIBONACCI[k] = Functions.FIBONACCI.z(k + 2);
     }
   }
