+++ src/irvine/oeis/a027/A027377.java
@@ -36,7 +36,7 @@ public class A027377 extends AbstractSequence {
     for (final Z dd : Jaguar.factor(mN).divisors()) {
       final int d = dd.intValue();
       final Z z = Z.ONE.shiftLeft(2L * d);
-      sum = ZUtils.mobiusAdd(Functions.MOBIUS.i((long) (mN / d)), sum, z);
+      sum = ZUtils.mobiusAdd(Functions.MOBIUS.i(mN / d), sum, z);
     }
     return sum.divide(mN);
   }
