+++ src/irvine/oeis/a034/A034743.java
@@ -18,7 +18,7 @@ public class A034743 extends Sequence1 {
     Z sum = Z.ZERO;
     for (final Z dd : Jaguar.factor(++mN).divisors()) {
       final int d = dd.intValue();
-      sum = sum.add(Functions.BELL.z(d - 1).multiply(Functions.MOBIUS.i((long) (mN / d))));
+      sum = sum.add(Functions.BELL.z(d - 1).multiply(Functions.MOBIUS.i(mN / d)));
     }
     return sum;
   }
