+++ src/irvine/oeis/a034/A034745.java
@@ -18,7 +18,7 @@ public class A034745 extends Sequence1 {
     Z sum = Z.ZERO;
     for (final Z dd : Jaguar.factor(++mN).divisors()) {
       final int d = dd.intValue();
-      sum = sum.add(Z.THREE.pow(d - 1).multiply(Functions.FIBONACCI.z((long) (mN / d))));
+      sum = sum.add(Z.THREE.pow(d - 1).multiply(Functions.FIBONACCI.z(mN / d)));
     }
     return sum;
   }
